requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-background-tasks==1.2.8
django-redis==5.4.0
django-seed==0.3.1
django-tenant-schemas==1.12.0
django-tenants==3.7.0
djangorestframework==3.15.2
Faker==33.1.0
idna==3.10
mailtrap==2.0.1
ordered-set==4.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
